fix(346): address all reviewer blockers (Brad/Rahul/Diego)
- tsup external: add @beakerstack/logger to marketing-email-edge so it
  is resolved via Deno import map, not bundled inline (Rahul/Diego/Brad)
- manifest: add logger to marketing-email-edge dependsOn; build order now
  logger → lifecycle-events → marketing-email-edge (Rahul/Brad)
- sync-edge-shared: WARN→ERROR for missing import map entry; increment
  errors counter so script fails on misconfiguration (Rahul/Brad)
- sync-edge-shared: compare mapped path against expectedRelPath; catches
  stale or wrong import map entries, not just missing ones (Brad)
- sync-edge-shared: export topoSort, parseManifest, validateImportMapEntries
  as pure functions; topoSort throws instead of process.exit so tests can
  assert.throws; isMain guard prevents build on import; tests now import
  real implementations instead of inlining copies (Brad)
- waitlist-capture, stripe-webhook: migrate enqueueMarketingEmail import
  from deleted _shared/marketingEmailQueue.ts to @beakerstack/marketing-email/edge (Rahul)
- CI test.yml: build-edge now triggers on run_unit == true so packages/**
  changes validate the edge manifest; supabase-tests adds build-edge to
  needs so generated files exist before integration tests start (Brad/Diego)
- tsconfig: add kit-sync and waitlist-ops to include so @beakerstack/*
  path mappings resolve in editor for the new function directories (Diego)
