+++ packages/marketing-email/tsup.config.marketing-edge.ts
@@ -8,5 +8,5 @@ export default defineConfig({
   splitting: false,
   clean: true,
   outDir: 'dist',
-  external: ['@beakerstack/lifecycle-events', '@supabase/supabase-js'],
+  external: ['@beakerstack/lifecycle-events', '@beakerstack/logger', '@supabase/supabase-js'],
 });
